+++ android/src/main/java/com/plugin/flutter/zsdk/ErrorCode.java
@@ -8,5 +8,6 @@ public enum ErrorCode
     SUCCESS,
     EXCEPTION,
     PRINTER_ERROR,
+    PRINTER_REBOOTED,
     UNKNOWN,
 }