+++ lib/src/enumerators/error_code.dart
@@ -3,6 +3,7 @@ enum ErrorCode {
   SUCCESS,
   EXCEPTION,
   PRINTER_ERROR,
+  PRINTER_REBOOTED,
   UNKNOWN,
 }